pyproject.toml
[project]
name = "testapi"
version = "0.1.0"
description = "Test API"
authors = [
    {name = "Artem",email = "bedintema@gmail.com"}
]
requires-python = ">=3.9, <=3.10"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django4-background-tasks (>=1.2.10,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
